Refactored Controller Structure
