+++ .idea/data_source_mapping.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="DataSourcePerFileMappings">
-    <file url="file://$APPLICATION_CONFIG_DIR$/consoles/db/2b805fc8-d915-467f-8539-cbd5c8a9811f/console.sql" value="2b805fc8-d915-467f-8539-cbd5c8a9811f" />
     <file url="file://$PROJECT_DIR$/src/main/java/at/fhtw/swen1/repository/SessionRepository.java" value="2b805fc8-d915-467f-8539-cbd5c8a9811f" />
   </component>
 </project>
